[IMP] forwardport: move limits tests into their own file
test_simple is starting to get pretty big and we'll need some more
tests for #213
